삭ㅔ 시 사용중인 상품삭제안되도록 막음

+++ product/infra/repository/product_repo.py
@@ -95,6 +95,9 @@ class ProductRepository(IProductRepository):
         with SessionLocal() as db:
             product = db.query(Product).filter(Product.id == product_id).first()
 
+            if product.use_flag == 1:
+                raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR, detail="사용 중인 제품은 삭제 할 수 없습니다.") 
+
             db.delete(product)
             db.commit()
